+++ tests/test_logic/test_complexity/test_cognitive/conftest.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 """
 Fixtures to make testing cognitive complexity easy.