+++ tests/test_visitors/test_ast/test_naming/test_naming_rules/test_private.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 from wemake_python_styleguide.violations.naming import PrivateNameViolation
 from wemake_python_styleguide.visitors.ast.naming import WrongNameVisitor