+++ wemake_python_styleguide/logic/naming/blacklists.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 from functools import lru_cache
 from typing import FrozenSet