+++ wemake_python_styleguide/logic/naming/logical.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 from typing import Iterable
 
 from wemake_python_styleguide import constants