+++ wemake_python_styleguide/logic/tree/exceptions.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 import ast
 from inspect import getmro
 from typing import List, Mapping, Optional, Tuple