+++ wemake_python_styleguide/violations/base.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 """
 Contains detailed technical information about :term:`violation` internals.