+++ wemake_python_styleguide/violations/best_practices.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 """
 These checks ensure that you follow the best practices.