+++ wemake_python_styleguide/visitors/ast/compares.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 import ast
 from typing import ClassVar, List, Optional, Sequence